llmRoute/pyproject.toml
[project]
name = "llmroute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "fastapi",
    "uvicorn",
    "supabase",
    "streamlit>=1.51.0",
    "google-generativeai",
    "python-dotenv",
]
